zammad-ai/pyproject.toml
[project]
name = "zammad-ai"
license = "MIT"
version = "0.1.0"
description = "GenAI-powered agent for Zammad"
readme = "README.md"
requires-python = "==3.13.9"
authors = [
    { name = "Fabian Reinold", email = "fabian.reinold@muenchen.de" },
    { name = "Paul Ilitz", email = "paul.ilitz@muenchen.de" },
]
maintainers = [
    { name = "Fabian Reinold", email = "fabian.reinold@muenchen.de" },
]

dependencies = [
    "aiofiles==24.1.0",
    "feedparser==6.0.12",
    "langchain==0.3.27",
    "langchain-core==0.3.78",
    "langchain-openai==0.3.34",
    "langfuse==3.10.1",
    "logtools==0.8.7",
    "openai==2.1.0",
    "faststream[cli,kafka]==0.6.3",
    "httpx==0.28.1",
    "cryptography==46.0.3",
    "pydantic-settings==2.12.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    "truststore==0.10.4",
    "pandas==2.3.3",
]

[dependency-groups]
dev = [
    "ipykernel==7.1.0",
    "jupyter==1.1.1",
    "pre-commit==4.5.0",
    "pytest-asyncio==1.3.0",
    "pytest==9.0.1",
    "ruff==0.14.7",
]

[tool.uv]
add-bounds = "exact"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["test"]
asyncio_mode = "auto"
